# Treat endpoints consistently in calculations